requirements.txt
discord.py== 2.1.0
py3== 3.11.0
